fix(api): split auth and runtime handlers
