+++ docs/internals/go-package-reference.md
@@ -4774,6 +4774,7 @@ _No package overview is documented._
 - [`func (c *PlatformClient) CreateAdapterSession(ctx context.Context, req AdapterSessionRequest) (AdapterSession, error)`](#cli-platform-api-func-c-platformclient-createadaptersession-ctx-context-context-req-adaptersessionrequest-adaptersession-error)
 - [`func (c *PlatformClient) CreateTeam(ctx context.Context, slug, name string) (Team, error)`](#cli-platform-api-func-c-platformclient-createteam-ctx-context-context-slug-name-string-team-error)
 - [`func (c *PlatformClient) CreateTeamUser(ctx context.Context, slug, email, password, role string) (TeamMembership, error)`](#cli-platform-api-func-c-platformclient-createteamuser-ctx-context-context-slug-email-password-role-string-teammembership-error)
+- [`func (c *PlatformClient) CreateUser(ctx context.Context, email, password, role string) (PlatformUser, error)`](#cli-platform-api-func-c-platformclient-createuser-ctx-context-context-email-password-role-string-platformuser-error)
 - [`func (c *PlatformClient) CurrentPrincipal(ctx context.Context) (Principal, error)`](#cli-platform-api-func-c-platformclient-currentprincipal-ctx-context-context-principal-error)
 - [`func (c *PlatformClient) DeleteGrant(ctx context.Context, namespace, name string) error`](#cli-platform-api-func-c-platformclient-deletegrant-ctx-context-context-namespace-name-string-error)
 - [`func (c *PlatformClient) DeleteRuntimeServer(ctx context.Context, namespace, name string) error`](#cli-platform-api-func-c-platformclient-deleteruntimeserver-ctx-context-context-namespace-name-string-error)
@@ -4794,6 +4795,7 @@ _No package overview is documented._
 - [`func (c *PlatformClient) RecordImagePublish(ctx context.Context, record ImagePublishRecord) error`](#cli-platform-api-func-c-platformclient-recordimagepublish-ctx-context-context-record-imagepublishrecord-error)
 - [`func (c *PlatformClient) UpsertTeamMember(ctx context.Context, slug, userID, role string) (TeamMembership, error)`](#cli-platform-api-func-c-platformclient-upsertteammember-ctx-context-context-slug-userid-role-string-teammembership-error)
 - [`func (c *PlatformClient) ValidateCredentials(ctx context.Context) error`](#cli-platform-api-func-c-platformclient-validatecredentials-ctx-context-context-error)
+- [`type PlatformUser struct`](#cli-platform-api-type-platformuser-struct)
 - [`type Principal struct`](#cli-platform-api-type-principal-struct)
 - [`type ServerListItem struct`](#cli-platform-api-type-serverlistitem-struct)
 - [`type Team struct`](#cli-platform-api-type-team-struct)
@@ -4951,6 +4953,12 @@ func (c *PlatformClient) CreateTeamUser(ctx context.Context, slug, email, passwo
 
 ```
 
+<a id="cli-platform-api-func-c-platformclient-createuser-ctx-context-context-email-password-role-string-platformuser-error"></a>
+```text
+func (c *PlatformClient) CreateUser(ctx context.Context, email, password, role string) (PlatformUser, error)
+
+```
+
 <a id="cli-platform-api-func-c-platformclient-currentprincipal-ctx-context-context-principal-error"></a>
 ```text
 func (c *PlatformClient) CurrentPrincipal(ctx context.Context) (Principal, error)
@@ -5070,6 +5078,17 @@ func (c *PlatformClient) UpsertTeamMember(ctx context.Context, slug, userID, rol
 <a id="cli-platform-api-func-c-platformclient-validatecredentials-ctx-context-context-error"></a>
 ```text
 func (c *PlatformClient) ValidateCredentials(ctx context.Context) error
+
+```
+
+<a id="cli-platform-api-type-platformuser-struct"></a>
+```text
+type PlatformUser struct {
+	ID        string `json:"id"`
+	Email     string `json:"email"`
+	Role      string `json:"role"`
+	Namespace string `json:"namespace,omitempty"`
+}
 
 ```
