+++ internal/cli/platformapi/client.go
@@ -615,6 +615,13 @@ type Team struct {
 	CreatedAt time.Time `json:"created_at"`
 }
 
+type PlatformUser struct {
+	ID        string `json:"id"`
+	Email     string `json:"email"`
+	Role      string `json:"role"`
+	Namespace string `json:"namespace,omitempty"`
+}
+
 type TeamMembership = platform.TeamMembership
 
 type teamsResponse struct {
@@ -633,14 +640,8 @@ type teamMembershipResponse struct {
 	Membership TeamMembership `json:"membership"`
 }
 
-type teamUserResponse struct {
-	User struct {
-		ID        string `json:"id"`
-		Email     string `json:"email"`
-		Role      string `json:"role"`
-		Namespace string `json:"namespace"`
-	} `json:"user"`
-	Membership TeamMembership `json:"membership"`
+type userResponse struct {
+	User PlatformUser `json:"user"`
 }
 
 type namespaceListItem struct {
@@ -840,6 +841,39 @@ func (c *PlatformClient) CreateTeam(ctx context.Context, slug, name string) (Tea
 	return out.Team, nil
 }
 
+func (c *PlatformClient) CreateUser(ctx context.Context, email, password, role string) (PlatformUser, error) {
+	payload := struct {
+		Email    string `json:"email"`
+		Password string `json:"password"`
+		Role     string `json:"role,omitempty"`
+	}{
+		Email:    strings.TrimSpace(email),
+		Password: password,
+		Role:     strings.TrimSpace(role),
+	}
+	js, err := json.Marshal(payload)
+	if err != nil {
+		return PlatformUser{}, err
+	}
+	resp, err := c.do(ctx, http.MethodPost, "/users", "", bytes.NewReader(js))
+	if err != nil {
+		return PlatformUser{}, err
+	}
+	defer resp.Body.Close()
+	b, err := readBody(resp.Body)
+	if err != nil {
+		return PlatformUser{}, err
+	}
+	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
+		return PlatformUser{}, httpAPIError(resp.StatusCode, b)
+	}
+	var out userResponse
+	if err := json.Unmarshal(b, &out); err != nil {
+		return PlatformUser{}, err
+	}
+	return out.User, nil
+}
+
 func (c *PlatformClient) ListTeamMembers(ctx context.Context, slug string) ([]TeamMembership, error) {
 	rel := "/runtime/teams/" + url.PathEscape(strings.TrimSpace(slug)) + "/members"
 	resp, err := c.do(ctx, http.MethodGet, rel, "", nil)
@@ -862,33 +896,16 @@ func (c *PlatformClient) ListTeamMembers(ctx context.Context, slug string) ([]Te
 }
 
 func (c *PlatformClient) CreateTeamUser(ctx context.Context, slug, email, password, role string) (TeamMembership, error) {
-	payload := map[string]string{
-		"email":    strings.TrimSpace(email),
-		"password": password,
-		"role":     strings.TrimSpace(role),
-	}
-	js, err := json.Marshal(payload)
+	user, err := c.CreateUser(ctx, email, password, "")
 	if err != nil {
 		return TeamMembership{}, err
 	}
-	rel := "/runtime/teams/" + url.PathEscape(strings.TrimSpace(slug)) + "/users"
-	resp, err := c.do(ctx, http.MethodPost, rel, "", bytes.NewReader(js))
+	membership, err := c.UpsertTeamMember(ctx, slug, user.ID, role)
 	if err != nil {
 		return TeamMembership{}, err
 	}
-	defer resp.Body.Close()
-	b, err := readBody(resp.Body)
-	if err != nil {
-		return TeamMembership{}, err
-	}
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return TeamMembership{}, httpAPIError(resp.StatusCode, b)
-	}
-	var out teamUserResponse
-	if err := json.Unmarshal(b, &out); err != nil {
-		return TeamMembership{}, err
-	}
-	return out.Membership, nil
+	membership.Email = user.Email
+	return membership, nil
 }
 
 func (c *PlatformClient) UpsertTeamMember(ctx context.Context, slug, userID, role string) (TeamMembership, error) {
