+++ services/api/internal/runtimeapi/session_handlers.go
@@ -1,0 +1,198 @@
+package runtimeapi
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"log"
+	"net/http"
+	"strings"
+	"time"
+
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	mcpv1alpha1 "mcp-runtime/api/v1alpha1"
+	sentinelaccess "mcp-runtime/pkg/access"
+	runtimeaccess "mcp-sentinel-api/internal/runtimeapi/access"
+)
+
+func validateSessionRequest(req *accessSessionRequest) error {
+	req.Name = strings.TrimSpace(req.Name)
+	req.Namespace = runtimeaccess.DefaultAccessNamespace(req.Namespace)
+	req.ServerRef.Name = sentinelaccess.ServerName(strings.TrimSpace(string(req.ServerRef.Name)))
+	req.ServerRef.Namespace = sentinelaccess.Namespace(strings.TrimSpace(string(req.ServerRef.Namespace)))
+	req.Subject.HumanID = sentinelaccess.HumanID(strings.TrimSpace(string(req.Subject.HumanID)))
+	req.Subject.AgentID = sentinelaccess.AgentID(strings.TrimSpace(string(req.Subject.AgentID)))
+	req.Subject.TeamID = sentinelaccess.TeamID(strings.TrimSpace(string(req.Subject.TeamID)))
+	req.PolicyVersion = runtimeaccess.DefaultPolicyVersion(req.PolicyVersion)
+	req.ConsentedTrust = runtimeaccess.NormalizeTrust(req.ConsentedTrust)
+	if err := sentinelaccess.ValidateResourceName("name", req.Name); err != nil {
+		return err
+	}
+	if err := sentinelaccess.ValidateResourceName("namespace", req.Namespace); err != nil {
+		return err
+	}
+	if err := sentinelaccess.ValidateResourceName("serverRef.name", string(req.ServerRef.Name)); err != nil {
+		return err
+	}
+	if err := sentinelaccess.ValidateOptionalResourceName("serverRef.namespace", string(req.ServerRef.Namespace)); err != nil {
+		return err
+	}
+	if req.Subject.HumanID == "" && req.Subject.AgentID == "" && req.Subject.TeamID == "" {
+		return errors.New("one of subject.humanID, subject.agentID, or subject.teamID is required")
+	}
+	if err := runtimeaccess.ValidateTeamIDValue("subject.teamID", string(req.Subject.TeamID)); err != nil {
+		return err
+	}
+	if req.ConsentedTrust != "" && !runtimeaccess.ValidTrust(req.ConsentedTrust) {
+		return errors.New("consentedTrust must be low, medium, or high")
+	}
+	return nil
+}
+
+func (s *RuntimeServer) handleRuntimeSessionList(w http.ResponseWriter, r *http.Request) {
+	if s.accessMgr == nil {
+		writeAPIError(w, http.StatusServiceUnavailable, "kubernetes not available")
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
+	defer cancel()
+
+	namespace, err := s.scopedNamespaceForPrincipal(r.Context(), r.URL.Query().Get("namespace"))
+	if err != nil {
+		writeAPIError(w, http.StatusForbidden, err.Error())
+		return
+	}
+	sessions, err := s.accessMgr.ListSessions(ctx, namespace)
+	if err != nil {
+		writeAPIError(w, http.StatusInternalServerError, "failed to list sessions")
+		return
+	}
+
+	p, filterByPrincipal := principalFromContext(ctx)
+	filterByPrincipal = filterByPrincipal && p.Role != roleAdmin
+	var serverCache accessServerCache
+	if filterByPrincipal {
+		serverCache, err = s.accessServerCacheForSessionRefs(ctx, namespace, sessions.Items)
+		if err != nil {
+			log.Printf("runtime session list: list MCPServers for visibility failed: %v", err)
+			writeAPIError(w, http.StatusInternalServerError, "failed to inspect server references")
+			return
+		}
+	}
+
+	summaries := make([]sentinelaccess.SessionSummary, 0, len(sessions.Items))
+	for _, sess := range sessions.Items {
+		if filterByPrincipal && !runtimeaccess.AccessRefVisibleWithServerCache(sess.Namespace, sess.Spec.ServerRef, serverCache,
+			func(server mcpv1alpha1.MCPServer) bool { return principalCanAdministerMCPServer(p, server) },
+			func(namespace string, serverLabels map[string]string) bool {
+				return principalCanAdministerServerLabels(p, namespace, serverLabels)
+			},
+		) {
+			continue
+		}
+		summaries = append(summaries, sentinelaccess.ToSessionSummary(sess))
+	}
+
+	writeJSON(w, http.StatusOK, map[string]interface{}{"sessions": summaries})
+}
+
+func (s *RuntimeServer) handleRuntimeSessionApply(w http.ResponseWriter, r *http.Request) {
+	if s.accessMgr == nil {
+		writeAPIError(w, http.StatusServiceUnavailable, "kubernetes not available")
+		return
+	}
+	if p, ok := principalFromContext(r.Context()); ok && p.Role != roleAdmin {
+		writeAPIError(w, http.StatusForbidden, "admin role required")
+		return
+	}
+
+	var req accessSessionRequest
+	r.Body = http.MaxBytesReader(w, r.Body, accessApplyMaxBytes)
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		writeBodyDecodeError(w, err)
+		return
+	}
+	if p, ok := principalFromContext(r.Context()); ok && p.Role != roleAdmin && strings.TrimSpace(req.Namespace) == "" {
+		req.Namespace = strings.TrimSpace(p.Namespace)
+	}
+	if err := validateSessionRequest(&req); err != nil {
+		writeAPIError(w, http.StatusBadRequest, err.Error())
+		return
+	}
+	scopedNamespace, err := s.scopedAccessWriteNamespaceForPrincipal(r.Context(), req.Namespace)
+	if err != nil {
+		writeAPIError(w, http.StatusForbidden, err.Error())
+		return
+	}
+	req.Namespace = scopedNamespace
+	if err := runtimeaccess.BindAccessServerRefNamespace(req.Namespace, &req.ServerRef); err != nil {
+		writeAPIError(w, http.StatusBadRequest, err.Error())
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
+	defer cancel()
+
+	targetServer, err := s.accessMgr.GetMCPServerRef(ctx, req.ServerRef)
+	if err != nil {
+		if sentinelaccess.IsMCPServerNotFoundForRef(err) {
+			writeAPIError(w, http.StatusBadRequest, err.Error())
+		} else {
+			log.Printf("runtime session: assert MCPServer ref failed: %v", err)
+			writeAPIError(w, http.StatusInternalServerError, "failed to verify server reference")
+		}
+		return
+	}
+	if err := s.bindAccessSubjectTeamID(ctx, req.Namespace, targetServer.Spec.TeamID, &req.Subject); err != nil {
+		writeAPIError(w, http.StatusForbidden, err.Error())
+		return
+	}
+	if !s.principalCanAdministerAccessServer(r.Context(), *targetServer) {
+		writeAPIError(w, http.StatusForbidden, "forbidden server")
+		return
+	}
+
+	revoked, err := s.sessionRevokedForApply(ctx, req)
+	if err != nil {
+		log.Printf("read session state %s/%s failed: %v", req.Namespace, req.Name, err)
+		writeAPIError(w, http.StatusInternalServerError, "failed to read session state")
+		return
+	}
+
+	session := &sentinelaccess.MCPAgentSession{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      req.Name,
+			Namespace: runtimeaccess.DefaultAccessNamespace(req.Namespace),
+		},
+		Spec: sentinelaccess.MCPAgentSessionSpec{
+			ServerRef:      req.ServerRef,
+			Subject:        req.Subject,
+			ConsentedTrust: req.ConsentedTrust,
+			ExpiresAt:      req.ExpiresAt,
+			Revoked:        revoked,
+			PolicyVersion:  runtimeaccess.DefaultPolicyVersion(req.PolicyVersion),
+		},
+	}
+	applied, err := s.accessMgr.ApplySession(ctx, session)
+	if err != nil {
+		writeK8sApplyError(w, "session", session.Namespace, session.Name, err)
+		return
+	}
+	writeJSON(w, http.StatusOK, map[string]interface{}{"session": sentinelaccess.ToSessionSummary(*applied)})
+}
+
+func (s *RuntimeServer) sessionRevokedForApply(ctx context.Context, req accessSessionRequest) (bool, error) {
+	if req.Revoked != nil {
+		return *req.Revoked, nil
+	}
+	existing, err := s.accessMgr.GetSession(ctx, req.Name, runtimeaccess.DefaultAccessNamespace(req.Namespace))
+	if apierrors.IsNotFound(err) {
+		return false, nil
+	}
+	if err != nil {
+		return false, err
+	}
+	return existing.Spec.Revoked, nil
+}
